Updated bot.py to v2
I have released this verison of the bot as I am no longer hosting the bot. This version was previously running on the live, hosted version of the bot and is the most up to date version. Feel free to fork, copy, or break this as much as you want :)